backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-CORS==4.0.1
marshmallow==3.21.3
Werkzeug==3.0.3
python-dotenv==1.0.1
pytest==8.2.2
pytest-flask==1.3.0
coverage==7.5.4
